example/requirements.txt
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
